# docs/api/openapi-fase1.yaml
openapi: 3.0.3
info:
  title: Fidelizar API — Phase 1 (Counter web)
  version: "1.0.0-f1"
  description: >
    F1-04b's REST contract. Hand-written — no OpenAPI generation package was added
    (CLAUDE.md: no new dependency without asking; a static document needs none).
    Endpoints marked `x-status: pending` are contract-only: no controller exists for them
    because no Application service backs them yet. See docs/REST-CONTRACT-F1.md for the
    endpoint-by-endpoint rationale and the service each pending one is waiting on.
servers:
  - url: /api
...
paths:
...
  /miembros:
    get:
      tags: [Miembros]
      summary: S2 Buscar socio — search by name (FUNCTIONAL-SPEC §4)
      x-status: implemented
      description: >
        Backed by IMiembroBusquedaService.BuscarAsync, which queries NombreNormalizado
        (any query word must appear as a substring — accent/case folding already baked
        into NombreNormalizado) and rejects a query under 3 normalised characters with
        400 BUSQUEDA_MUY_CORTA rather than falling back to a listing. There is
        deliberately no "list all members" variant of this endpoint for any role below
        Dueño (FUNCTIONAL-SPEC §4, ARCHITECTURE §8).
      security: [{ cookieAuth: [] }]
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string, minLength: 3 }
          description: Cashiers search; there is no browsing (FUNCTIONAL-SPEC §4).
      responses:
        "200":
          description: OK — candidates, each with enough to disambiguate homonyms, never phone/DNI.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/MiembroResumen' }
        "401":
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Miembros]
      summary: S5 Alta de socio (FUNCTIONAL-SPEC §7)
      x-status: pending
      x-pending-service: >
        Needs an Application use case that creates the Miembro and, in the same
        transaction, the mandatory DatosPersonales Consentimiento row (I10) — and the
        optional DatosSensibles one that gates diet/allergy fields. IMiembroRepository
        only has AddAsync today; nothing composes it with Consentimiento.
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AltaMiembroRequest' }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/MiembroResumen' }
        "400":
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'

  # ---------------------------------------------------------------------
  # S3 · Ficha del socio — partially implemented
  # ---------------------------------------------------------------------
  /miembros/{miembroId}/saldo:
    get:
      tags: [Miembros]
      summary: S3 Ficha del socio — balance and cutoff date only
      x-status: implemented
      description: >
        Backed by ISaldoService + ICorteService. ISaldoService.ObtenerSaldoAsync looks the
        member up before computing SUM(Monto), so an unknown miembroId — or one that belongs
        to a different Negocio — 404s instead of returning a misleading 0.
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/MiembroId'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SaldoMiembroResponse' }
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          description: No member with this id exists for the caller's Negocio.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        "409":
          description: No Corte declared for this Negocio yet (CORTE_NO_DECLARADO).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /miembros/{miembroId}/ficha-mostrador:
    get:
      tags: [Miembros]
      summary: S3 Ficha del socio — the full counter view (name, number, alerts)
      x-status: implemented
      description: >
        Backed by IFichaMostradorService, which 404s an unknown/another-business's
        miembroId (IMiembroRepository.GetByIdAsync) and builds the alert strip. Phase 1
        can only ever populate the Cumpleanos alert (RN-11: notice from 2 days before,
        day/month only) — AlergiaODieta needs PerfilMiembro (phase 3, F3-01) and
        ComprasHabituales needs phase-4 aggregation, so neither is ever produced yet.
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/MiembroId'
      responses:
        "200":
          description: OK — never phone or DNI for a Cajero (FUNCTIONAL-SPEC §5).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FichaMostradorResponse' }
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          description: No member with this id exists for the caller's Negocio.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /miembros/{miembroId}/completo:
    get:
      tags: [Miembros]
      summary: S6 Ficha completa (Encargada/Dueño only)
      x-status: implemented
      description: >
        Backed by IFichaCompletaService — the same Miembro lookup as ficha-mostrador,
        plus phone/DNI, plus a RegistroAuditoria row ("VerFichaCompleta", DATA-MODEL §2)
        written on every single read, not only the first.
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/MiembroId'
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/FichaCompletaResponse' }
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          description: No member with this id exists for the caller's Negocio.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

  # ---------------------------------------------------------------------
  # S4 · Registrar canje — implemented
  # ---------------------------------------------------------------------
  /miembros/{miembroId}/canjes:
    post:
      tags: [Movimientos]
      summary: S4 Registrar canje (FUNCTIONAL-SPEC §6)
      x-status: implemented
      description: >
        Idempotent on claveIdempotencia (README decision #6, 2026-08-19). Backed by
        ISaldoService.RegistrarCanjeAsync, which checks IMovimientoRepository.GetPorClaveIdempotenciaAsync
        before writing anything: a retry with the same key and the same monto/fechaEfectiva/motivo
        returns the original 200 CanjeResponse, no second row. The actual guarantee against two
        concurrent identical requests both writing a Canje lives in a unique partial index on
        (NegocioId, ClaveIdempotencia), not only in this check.
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/MiembroId'
        - $ref: '#/components/parameters/CsrfHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RegistrarCanjeRequest' }
      responses:
        "200":
          description: >
            Redemption registered — the new balance, shown immediately. Also returned, with the
            same body, when claveIdempotencia matches an earlier successful attempt.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CanjeResponse' }
        "400":
          description: >
            MONTO_INVALIDO (≤ 0), CANJE_SUPERA_SALDO (I6/RN-24), SALDO_EN_REVISION (RN-25) or
            CLAVE_IDEMPOTENCIA_REQUERIDA (missing/blank claveIdempotencia).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        "401":
          $ref: '#/components/responses/Unauthorized'
        "409":
          description: >
            CLAVE_IDEMPOTENCIA_REUTILIZADA — claveIdempotencia was already used for a different
            member, amount, date or reason. Not a retry; a client bug.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

  # ---------------------------------------------------------------------
  # S7 · Historial de movimientos — implemented
  # ---------------------------------------------------------------------
  /miembros/{miembroId}/movimientos:
    get:
      tags: [Movimientos]
      summary: S7 Historial de movimientos (Encargada/Dueño only)
      x-status: implemented
      description: >
        Backed by IHistorialMovimientosService, a thin wrapper over
        IMovimientoRepository.GetPorMiembroAsync that also resolves each movement's
        UsuarioNombre and 404s an unknown/another-business's miembroId.
      security: [{ cookieAuth: [] }]
      parameters:
        - $ref: '#/components/parameters/MiembroId'
      responses:
        "200":
          description: OK — newest first.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/MovimientoResponse' }
        "401":
          $ref: '#/components/responses/Unauthorized'
        "404":
          description: No member with this id exists for the caller's Negocio.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

  # ---------------------------------------------------------------------
  # S8 · Anular movimiento — implemented
  # ---------------------------------------------------------------------
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: fidelizar_auth
      description: >
        JWT in an HttpOnly, Secure, SameSite=Strict cookie (ARCHITECTURE §8). Every
        state-changing request additionally requires the X-CSRF-TOKEN header.
  parameters:
    MiembroId:
      name: miembroId
      in: path
      required: true
      schema: { type: integer }
    CsrfHeader:
      name: X-CSRF-TOKEN
      in: header
      required: true
      schema: { type: string }
      description: From GET /auth/csrf-token. Required on every state-changing request (ARCHITECTURE §8).
  responses:
    Unauthorized:
      description: No valid session, or the role lacks the required policy.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    BadRequest:
      description: Validation failure — see `details` for the offending field(s).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [errorCode, message]
      properties:
        errorCode: { type: string, example: CANJE_SUPERA_SALDO }
        message: { type: string, example: "El canje ($150.00) supera el saldo disponible ($100.00)." }
        details:
          type: array
          items:
            type: object
            properties:
              field: { type: string }
              issue: { type: string }

    # ---- Implemented (Fidelizar.Shared) ----
...
    SaldoMiembroResponse:
      type: object
      required: [miembroId, saldo, corteFecha]
      properties:
        miembroId: { type: integer }
        saldo: { type: number, format: double, description: "decimal on the wire; 2 decimals (I4)" }
        corteFecha: { type: string, format: date }
    RegistrarCanjeRequest:
      type: object
      required: [monto, fechaEfectiva, motivo, claveIdempotencia]
      properties:
        monto: { type: number, format: double, minimum: 0.01 }
        fechaEfectiva: { type: string, format: date, description: "May be in the past (ARCHITECTURE §10)." }
        motivo:
          type: string
          description: >
            Marked required here because MovimientoCredito.Crear currently requires it for
            every Canje regardless of date. FUNCTIONAL-SPEC §6 describes it as optional when
            fechaEfectiva is today — a real discrepancy between the two documents, flagged
            in docs/REST-CONTRACT-F1.md, not resolved by this contract.
        claveIdempotencia:
          type: string
          maxLength: 100
          description: >
            Generated by the client, one per redemption attempt — not one per logical
            redemption (README decision #6, 2026-08-19). A retry after a lost response
            (FLOW-S2-S5.md §3.8) reuses the same value: the server recognises it and returns
            the original CanjeResponse instead of writing a second Canje. The same key with
            different monto/fechaEfectiva/motivo/member is rejected as CLAVE_IDEMPOTENCIA_REUTILIZADA
            (409) — that is not a retry, it is a client mistake.
    CanjeResponse:
      type: object
      required: [movimientoId, monto, fechaEfectiva, saldoResultante]
      properties:
        movimientoId: { type: integer, format: int64 }
        monto: { type: number, format: double }
        fechaEfectiva: { type: string, format: date }
        saldoResultante: { type: number, format: double, description: "Historical evidence only (I2), never re-derived from this." }

    # ---- Pending — documented shape, not yet a C# type in Shared ----
    MiembroResumen:
      type: object
      description: >
        S2 search result row — name, member number, balance. Nothing else
        (FUNCTIONAL-SPEC §4). corteFecha closes the gap FLOW-S2-S5.md §1.4 flagged: R3
        ("not in any screen") requires the stale-data badge next to every rendered
        balance, including a search result row.
      properties:
        id: { type: integer }
        nombre: { type: string }
        numeroSocio: { type: string, nullable: true }
        saldo: { type: number, format: double }
        fechaAlta: { type: string, format: date, description: Shown only to disambiguate homonyms. }
        corteFecha: { type: string, format: date }
    FichaMostradorResponse:
      type: object
      description: S3 full counter view — never phone or DNI for a Cajero.
      properties:
        id: { type: integer }
        nombre: { type: string }
        numeroSocio: { type: string, nullable: true }
        saldo: { type: number, format: double }
        corteFecha: { type: string, format: date }
        alertas:
          type: array
          items:
            type: object
            properties:
              tipo: { type: string, enum: [SaldoDisponible, Cumpleanos, AlergiaODieta, ComprasHabituales] }
              texto: { type: string }
    FichaCompletaResponse:
      type: object
      description: S6 — Encargada/Dueño only.
      properties:
        id: { type: integer }
        nombre: { type: string }
        numeroSocio: { type: string, nullable: true }
        clienteExternoId: { type: string, nullable: true }
        telefono: { type: string, nullable: true }
        dni: { type: string, nullable: true }
        fechaNacimiento: { type: string, nullable: true, description: "Day and month only, RN-11." }
        sucursalId: { type: integer, nullable: true }
        activo: { type: boolean }
    AltaMiembroRequest:
      type: object
      description: S5 — consent is part of the form, not an afterthought (I10).
      required: [nombre, consentimientoDatosPersonales]
      properties:
        nombre: { type: string }
        clienteExternoId: { type: string, nullable: true }
        telefono: { type: string, nullable: true }
        dni: { type: string, nullable: true }
        fechaNacimientoDia: { type: integer, nullable: true }
        fechaNacimientoMes: { type: integer, nullable: true }
        sucursalId: { type: integer, nullable: true }
        consentimientoDatosPersonales: { type: boolean, description: "Must be true — explicit checkbox." }
        consentimientoDatosSensibles: { type: boolean, description: "Without it, diet/allergy fields are not shown at all (I10)." }
    MovimientoResponse:
      type: object
      description: S7/S8 — one ledger row.
      properties:
        id: { type: integer, format: int64 }
        tipo: { type: string, enum: [SaldoInicial, Acumulacion, Canje, Ajuste] }
        monto: { type: number, format: double }
        fechaEfectiva: { type: string, format: date }
        registradoEn: { type: string, format: date-time }
        motivo: { type: string, nullable: true }
        saldoResultante: { type: number, format: double }
        usuarioNombre: { type: string, nullable: true }
...
security:
  - cookieAuth: []